AG-12847 - controls column module
