+++ community-modules/core/src/grid.ts
@@ -14,7 +14,6 @@ import { ColumnMoveService } from './columns/columnMoveService';
 import { ColumnNameService } from './columns/columnNameService';
 import { ColumnSizeService } from './columns/columnSizeService';
 import { ColumnViewportService } from './columns/columnViewportService';
-import { ControlsColService } from './columns/controlsColService';
 import { FuncColsService } from './columns/funcColsService';
 import { PivotResultColsService } from './columns/pivotResultColsService';
 import { VisibleColsService } from './columns/visibleColsService';
@@ -511,7 +510,6 @@ export class GridCoreCreator {
             ColumnNameService,
             ColumnViewportService,
             PivotResultColsService,
-            ControlsColService,
         ];
 
         const moduleBeans = this.extractModuleEntity(rowModelModules, (module) => (module.beans ? module.beans : []));
