+++ community-modules/core/src/gridCoreModule.ts
@@ -1,6 +1,6 @@
 import { AlignedGridsModule } from './alignedGridsModule';
 import { CommunityApiModule } from './api/apiModule';
-import { ColumnAutosizeModule, DataTypeModule } from './columns/columnModule';
+import { ColumnAutosizeModule, ControlsColumnModule, DataTypeModule } from './columns/columnModule';
 import { EditModule } from './edit/editModule';
 import { FilterModule } from './filter/filterModule';
 import { _defineModule } from './interfaces/iModule';
@@ -35,5 +35,6 @@ export const CommunityFeaturesModule = _defineModule({
         PaginationModule,
         CommunityApiModule,
         ColumnAutosizeModule,
+        ControlsColumnModule,
     ],
 });
